[RENAME/#20] 유저 일정 패키지명 변경
