+++ src/main/java/com/togedy/togedy_server_v2/domain/schedule/entity/Category.java
@@ -1,6 +1,6 @@
-package com.togedy.togedy_server_v2.domain.calendar.entity;
+package com.togedy.togedy_server_v2.domain.schedule.entity;
 
-import com.togedy.togedy_server_v2.domain.calendar.dto.PatchCategoryRequest;
+import com.togedy.togedy_server_v2.domain.schedule.dto.PatchCategoryRequest;
 import com.togedy.togedy_server_v2.domain.user.entity.User;
 import com.togedy.togedy_server_v2.global.entity.BaseEntity;
 import com.togedy.togedy_server_v2.global.entity.Status;
